🔧 adding configuration for run experiments

+++ run_transfer_single.sh
@@ -8,23 +8,34 @@
     # Transfer to an obstacle course environment
 
 echo "Cleaning old models and logs.."
-rm -rf models
-rm -rf logs
+
+#!/bin/bash
+
+read -p "Are you sure you want to delete 'models' and 'logs'? (y/n): " answer
+
+if [ "$answer" = "y" ] || [ "$answer" = "Y" ]; then
+    rm -rf models
+    rm -rf logs
+    echo "Directories deleted."
+else
+    echo "Operation cancelled."
+fi
 
 mkdir logs
 
-echo "Train PPO (1/3)"
+echo "Train and Test PPO (1/3)"
 sh scripts/ppo/run_train.sh
-echo "Train PPO (1/3) 👌"
+sh scripts/ppo/run_test.sh
+echo "Train and Test PPO (1/3) 👌"
 
-echo "Train PPO Curriculum (2/3)"
-sh scripts/ppo/run_train_curriculum.sh
-echo "Train PPO Curriculum (2/3) 👌"
+#echo "Train PPO Curriculum (2/3)"
+#sh scripts/ppo/run_train_curriculum.sh
+# echo "Train PPO Curriculum (2/3) 👌"
 
-echo "Train Plastic-transfer(3/3)"
-sh scripts/plastic-transfer/run_train.sh
-echo "Train Plastic-transfer(3/3) 👌"
+# echo "Train Plastic-transfer(3/3)"
+#sh scripts/plastic-transfer/run_train.sh
+# echo "Train Plastic-transfer(3/3) 👌"
 
-# sh scripts/ppo/run_test.sh
+
 # sh scripts/ppo/run_test_curriculum.sh
 # sh scripts/plastic-transfer/run_test.sh
